Fix --download-skip and enhance GUID file.
The GUID file was enhanced to note the episode titles along with the
episode GUIDs for easier reference.  In addition, when GUIDs are
written to the file, the time when the updates finished is added as a
comment, also for reference.

Also, --download-skip was fixed to actually skip downloading
episodes instead of just skipping the writing of episodes.